Extract logger in separate crate and expose Logger trait
Instead of providing a hard coded logger instance ourselves we now just
export a new Logger trait that conforms to the way we expect loggers to
write request logs. For now this should not make a big difference as the
trait is not quite generic yet, but this should work as a necessary step
for sufficient efforts in accepting a more generic and customizable
logging approach.